Beginning of 2.0.0 Rewrite
Re-Writing the module to be Object Oriented and more pythonic.

This commit holds the initiale KipptAPI object which allows you to authenticate with either
a user/pass or user/token combo. If a password is used to authenticate we immediately grab the
API token instead so as not to be passing the password around.

The KipptAPI can then call any other object, for example - The only one written in this commit -
the clips object; and of course a clips object has many clips.

Example use:

k = KipptAPI('my_username', 'my_password')

k.clips().all(limit=5)

k.clips().clip(12345)

k.clips().clip(12345).delete()